Implement toggleable pinning for social sidebar.
This will only be available above a certain width (below that width,
this is always an overlay regardless of what is stored in local
storage). When in overlay mode, performing any navigation will close the
overlay (this is a bit brittle, but we don't add a ton of navigation to
this in practice. If new navigation is added, make sure to call
`onNavigation` from `useNavigationTracker`).

No animations for the overlay mode yet, but I need to add them to the
navigation menu and it's like the same implementation, so I'll do them
at the same time.