Remove arrow from tooltips, adjust tooltip margin to 4px.
Fixes #1125. This matches the more modern Material style, I don't think
the arrows were really adding much, and the arrow was pretty
confusing/misleading when at the edge of the screen. We could have
potentially moved the arrow but there's a lot of complexity in doing
that, especially given I don't think it really adds a ton.