Add a scalar + newtype for user IDs.
I avoided doing this before because it's a bit annoying with the sqlx
query macros (you have to `as _` the field a lot of the time or
potentially map it to the inner type). It seems like they may fix this
issue in the near future, but for now I think it's worthwhile on the
graphql side anyway.