Don't sort the users in the user entries selector.
Not sure why we were sorting the users here when we specifically had a
separate function to sort the users which we would call separately in
all of our usages. Had to remove it now so we can display the last 10
chatters in the order they posted, and not alphabetically.