Refactor the game players display to use flexbox instead of CSS columns.
This has the benefit of being able to use gaps instead of margins, and
also not having to insert fake elements to keep the teams ordered.

This commit also changes the order in which the teams/players are
ordered so they always keep the consistent order. Basically the ordering
follows two rules:
- Put the team/player of the user whose profile we're currently viewing
  on the left side
- Order the players within team alphabetically