Move email verification to a purely code-based system, with more persistent encouragement.
This is similar to the earlier password reset changes, with the
intention of using codes that are easy to copy-paste from email directly
into the client. Email verification is now done via a dialog instead of
being a separate page you can navigate to. This dialog can be opened
from account settings, or it will open automatically on logins after the
initial signup (provided the account has an unverified email).

Integration tests that don't care about email verification will have to
suppress this dialog if they don't want to deal with it (a utility for
this has been provided).