Add jotai as a redux replacement for starcraft health state.
Jotai seems nice for removing a lot of the indirection that exists
through redux (although I could see redux being useful for *some* things
going forward, for stuff that is very independent from the rest of the
app, it's nice not to have to construct action types + write code to
store the payloads). We can generally mix these and choose the right
tool for the job.

The StarCraft health stuff seemed like a nice testing ground because it
is small and self-contained.

This includes jotai-devtools, although currently it is useless because
it doesn't seem to work with custom stores properly. I've filed an
issue, so hopefully that gets fixed soon.