Handle offline state for play content.
I think this is sufficient offline handling for the app for now
(everything else might not function well or might do weird things if you
manage to be able to send HTTP requests but not have a websocket
connection, but play stuff is the only stuff that requires a WS
connection to be able to make the normal HTTP requests).

So yeah, probably think about a general UI design for offline state
handling for other stuff and improve it later.