Hide the map favorite button if the user is not logged in. (#1173)
Initially I thought we should still show the button, but redirect them
to the login page if they click it. But then I thought would the user
expect for the map to be favorited once they log in? Unsure, so this
seemed simpler :d

Also, I've removed the favoriting option from the lobbies. It seemed
kind of weird (and half-broken) being able to favorite other people's
private maps.