Update the local visions variable as well when changing replay visions
This variable is also used during replays, at least when determining when
to create fog sprites.

There would be also a 'local exploration visions' variable, which BW
usually keeps up to date too when changing replay visions, but we don't
have analysis available for it, so hoping for now that we can get away
without anything breaking even if it doesn't get updated properly on
replay vision changes.

Fixes #1169