Implement a gameplay activity widget.
This is a replacement for changing the Play button's content +
destination, instead we will show a floating (and draggable) widget over
top of everything for matchmaking + lobby status. For now I didn't build
an active game one because it didn't feel particularly useful to be able
to get back to that page, but if it seems worthwhile it'd be easy to
add.

There's still some things that would be useful to have in these widgets
(e.g. showing match acceptance here and hiding the dialog after accept,
countdown for lobbies, etc.). I'll file issues for these. Animation also
isn't there yet, but I have an idea for that and I'll implement it in a
follow-up.

As part of this, I moved settings z-index higher so it displays over the
widget.