Change useObservedDimensions/useResizeObserver to default to border-box sizing.
The previous default was to use `contentRect`, but according to MDN this
is deprecated and could be removed in a future standards update. It also
just wasn't generally what we wanted, since it excludes border+padding
from the given size. If that's what you want still, you can now pass
options to `useObservedDimensions` to get the content size instead.