Adjust transitions for play button.
This adds a transition for clicks (since that felt weirdly absent
before) that does a slight scale (this felt nicer to me than trying to
mess with the gradient circles, and also avoided needing to make yet
another related transition on the same property).

Still not totally happy with this, but I'm not quite sure how to get
motion to do this well, having a time-based animation that runs and
repeats but can also be made interactive in certain moments seems not
great (and I especially dislike that `null` keyframes get used when
repeating animations, I'd prefer they just get used initially and then
dropped =/ ). There might be some better way of going about this, but I
tried a few things and what we have now was the best I got.

The motion of the gradient circles is also just a bit unsatisfying, but
doesn't seem worth spending a ton more time on it currently.