Refactor game loader to handle notifying clients of routes on its own, rather than having each service do it itself.
The handling in matchmaking and lobbies for this was the same, just with
different (identical) websocket messages, so it was easy to simply pull
it into one spot. I'll also be doing this sort of refactor for game
setup, but that is intertwined with more things in lobbies/matchmaking
so it will be a bit more work to unify.