Remove 5 second 'map selection' timer from matchmaking.
This is an obstacle to having the game loader handle all of the game
setup process (as lobbies don't have this). Ideally we'd still guarantee
that the map, etc. are shown for X seconds in both modes, but I think
this will be a lot easier to do when we move to ingame loading screens
(which should happen Soon (TM)). For now this is maybe a slight
degradation, but I'll see what I can clean up once this refactor is
completed.