Refactor game loader to handle all the game setup instead of calling out to its callers to send messages.
This commit breaks some of the user experience here, namely that the
matchmaking loading screen + countdown don't work. I'm going to look
into making an ingame loading screen next, if that doesn't work out
(because it's too much work for now) then I'll reimplement a matchmaking
loading screen in the app (we needed to convert it to TS anyway so it's
kind of the same amount of work).