Do initial work for ingame loading screen.
This changes things so that the game window is never hidden once
launched, with the plan that we can instead render to it via egui during
the init period.

This also adds window position/size restoration for windowed mode
(fullscreen modes need more work unfortunately).

As of this commit no rendering happens after the game init hook is
reached until we are ingame, working on figuring out how to trigger
rendering in that middle area.