Implement the beginnings of an ingame loading screen.
With this approach, we no longer need a loading screen in the app,
instead we simply launch the game and it displays a window immediately,
where we can display the map image, match info, and loading state to
users, then do a countdown before starting the game. This should resolve
issues with people whose computers take a bit to do the very last
loading step, not lose the first few seconds on the game on streams, and
play a lot more nicely with exclusive fullscreen mode. We generally just
need less hooks into Windows APIs now as the game can mostly manage its
window normally.

The loading screen is rendered via egui, although as of this commit it
is not fully completed yet. It's in a working state though, and I don't
want this branch to get too far out of sync with the main branch since
it has some expansive changes, so I'm merging it now.

This commit also:

- adds window position/size restoration for windowed mode
  (fullscreen modes need more work unfortunately).
- adds a way to play sounds ingame
- adds a wrapper around BwString to make it safer to use
- adds support for printing text messages (both as normal chat and
  centered)
- adds a way to handle text commands (currently just one: /version).
- sets scr-only flag for bw_dat so we only get theh correct structures
- fixes a crash if bw_vector_push is called on a 0 capacity vector
- adds a newtype for SbUserIds in game,
- fixes typings for rally-point-creator value that was a number and not a string.
- specifies image formats directly since avif's added deps seem to not build
  correctly on x64 (and we don't need that anyway).

Notes on problems that may be seen in the future:

I've noticed that the first frame we render the loading screen,
egui will render a warning (in debug mode) about request_discard being
called 3 frames in a row. I believe this is just because of the
surface/window resizing repeatedly during initialization, and can safely
be ignored. (It doesn't show up in release mode anyway)