Fix ClipCursor with always shown window, remove some now-unnecessary hooks, fix loading screen draw layer in release mode so it still hides FPS counter.
Also I've noticed that the first frame we render the loading screen,
egui will render a warning (in debug mode) about request_discard being
called 3 frames in a row. I believe this is just because of the
surface/window resizing repeatedly during initialization, and can safely
be ignored. (It doesn't show up in release mode anyway)