Implement basic loading screen layout in egui.
As of now this is missing user avatars and any indication of their
connectedness/loading state. I'll file issues for these as they probably
won't make it into the next release.

There are also some potential improvements that could be made to the
layout of users for various team game types.

This is implemented via an egui_extras Strip and then egui_flex layouts
inside each strip (I couldn't get egui_flex to size the player areas
evenly with uneven length names). I'm not sure that the flex stuff is
really necessary to build this, it could probably be done withh just
Strip + a normal egui layout, but I don't want to fight with it too much
right now.