Add countdown to ingame loading screen.
Getting the beeps to play at the right time (and in the right number)
required a number of hacks. At some point we should probably reverse
that code more deeply so it runs a bit more staightforwardly (I'll file
an issue for this).

Positioning/styling of the countdown still needs work, I just want to
get a version of this that works checked in in case I break it going
forward :)