Use typescript-result to more safely deal with GameLoader errors, utilize in lobbies for determining blame (blame determining is not good as of this commit, see TODOs).
In general I think we should probably reserve this approach for services
that are likely to be called by other services (and especially if there
is an A -> B -> C relationship there), as it is mainly useful to ensure
that callers have explicitly handled (or at least considered) all the
possible error types a function can return. It's a bit clunkier than in
e.g. Rust since the language doesn't provide any affordances for using
Results.