Add a more robust way of overriding compat flags people might have set on their StarCraft exe's.
We've seen a recent issue with someone's process requiring admin to
launch, which I assume is from compat settings, and whatever compat
settings are applied seem to be applied alongside `__COMPAT_LAYER`
anyway. This new approach just checks the registry key for these (which
is *generally* done on a user-by-user basis so it doesn't require admin
to change). If we find the key, we override it with a blank value before
launching the process, then restore the old value immediately
afterwards.