Turn off lint rule that requires values for boolean props.
I think I had originally turned this on because that seemed like the
direction JSX was moving in, but that definitely hasn't been the case
and I don't really have a preference for either direction, and a lot of
examples/things (e.g. motion) use naked boolean props so, seems nicer
for autocomplete.