Implement matchmaking bans for people that don't ready up or fail to load into a match.
Matchmaking bans are identifier-based, similar to account bans, although
when enacting a ban we only use the current active identifiers.

Bans are a ramping punishment system. The first level is a warning (no
actual punishment), the next level is 15 minutes, then next level is 1
hour, etc. After your ban expires, there is a certain duration
(depending on what level you reached) before it "clears". Until it
clears, any further bans continue to increase the level from that point.
Once all outstanding bans have cleared, you are reset to the initial
state (e.g. the next punishment will be a warning again).

We may want to use a system where it decays down one level at that
point, although it's a bit more complex to deal with so I implemented
the simpler thing now and we can see how it affects player behavior
first.

As of this commit, we don't notify players when a ban happens, we just
show them they're banned if they try to queue for a match. This isn't
ideal (especially given the warning-only level), but the notification
needs some more thought and isn't a super high priority right now. I'll
file an issue so we can pick up that work soon.