Add pre-match draft for team modes.
In team modes, this is a short time period where players select their
race in turn, letting players communicate with their teammates (and see
the map + opponent selections) while selecting. Opponents are anonymized
during this period (until the loading screen that follows). Leaving
during the draft has similar punishment to leaving while accepting or
during the game load.

This also cleans up some of the matchmaking-service code in general
(there's still a lot of party-related stuff here though that could be
simplified now, but I didn't want to do it while doing this work).

Client draft state is done in Jotai as Redux didn't feel particularly
useful for it. I've added a couple jotai extensions to make that more
feasible (one that lets us change the value of an atom when another atom
changes, and one that lets us have atoms that use immer for setting).

Some of the animations here are a little convoluted (namely the one to
flash the border of the racepicker) but I think they turned out decently
well.

This commit also:

- Adds a class for fadeable sounds to make fading them more DRY
- Fixes issues with clients disconnecting but the server not realizing
  for 30+ seconds (only a dev issue I am 99% certain, but it's fixed
  there now)
- Removes unused countdown stuff from matchmaking
- Adds a utility for trapping a user within a particular navigation state
  (basically FocusTrap but for navigating).