Fix desyncs in multiplayer games due to a sleep between do_lobby_game_init and try_finish_lobby_game_init.
Delays in that exact spot are unexpected by the game, as it expects to
be ticking the lobby state forward on each turn (and we were receiving
turns during the delay and not doing that).

I've moved some of this init code around so that the last bit of it
(do_lobby_game_init and try_finish_lobby_game_init) happens on the game
thread instead, and we have forge bring the window to the front directly
before that (seemed nicer to not have the window pop up a split second
*after* the game started if you had it minimized). This seems like it
will make things a bit smoother in general, since we'll no longer be
passing control between threads just as the game is starting.

This did require a bit of rearchitecting, since before we're running the
game loop we need to be processing Windows messages (and thus not
blocked), and Forge's window message pump has been stopped. I wrote a
little framepacing-esque code (albeit not that sophisticated and just
for receiving turns, not really for rendering) to try to hit the 42ms
turn times as closely as possible but not busy loop.