Add a way to zoom in and pan around map image preview.
Not a perfect implementation by all means, but might be enough for now?

Things that could be improved:
- add pinch/touch gestures for mobile support
- make the pan speed smoother on all zoom levels
- add the "slide" effect when panning fast and releasing a mouse
  (similar to what happens on phones / google maps)
- make the image quality even better than 2048x2048 on the highest zoom
  level?
- pull this logic into a common hook / control component that can be
  used on any element? (might come in handy if/when we add tournament
  brackets)

Also fixes one issue with buttons where if you release a mouse on a
disabled button, a `mouseup` event on the document would not run (this
conflicted with the slider), and adds a prop to render the slider
component without a balloon that shows the value.