Add restrictions feature for limiting account capabilities for a set time, use for chat restrictions.
This is a more general feature than just chat, but for now that's the
only one we have. This is very similar to the ban implementation, and
applies both on accounts and based on identifier matches.

I came up with a number of preset reasons, but we can add more if we
think of them (and will probably need more when we have more restriction
types).

The way we handle clearing restrictions on the client is maybe a bit
dumb right now, I think it will generally work alright but if the client
has an out-of-sync clock it may fall over (in that it may let them try
to send messages and error every time, not that it will let them send
messages in actuality). I have a TODO about a potential fix but we can
clean it up later.

As of this commit, this doesn't apply ingame, but I'll work on that
next.

This also fixes some random issues I noticed in the existing systems
(user_identifiers lacked a FK on user_id which allowed my dev server to
insert a bad row because of other bugs and messed up later queries until
I removed the bad data, user_bans was still using UUID v4 instead of the
newer stuff by default).