Add pointer-events: none; to all disabled buttons.
This fixes an issue where releasing a mouse on a disabled button would
not run the `mouseup` event on the `document`. An easy way where this
could have happened is when dragging a slider and then releasing a
slider on top of a disabled button, the slider would keep being dragged.