Add a lock around event processing during game init to prevent ClientSdk-related crash.
This uses a pretty dumb/basic spinlock implementation as it is easy to
use on multiple threads and doesn't use RAII for the unlocking. This
makes it somewhat inherently unsafe, but I think for this particular use
case that shouldn't matter much. I have seen it spin a bit long (1-10ms)
but since games are processor-hungry anyway and it's only during game
init, it doesn't seem that problematic.

We unlock the lock during resize/context menu and re-lock it when that
completes, since Windows will have prevented process_events from running
during that time as well (and that lets us keep stepping the lobby init
during that time).

There may be other paths that also need to be covered by this locks, but
these seem to be the main culprit for crashes right now.