Do our own message dispatch until we have finished modifying with the game state off-thread.
This seemed easier than trying to properly lock process_events out
across all the places we modified/accessed things. Should fix most
(all?) of the rest of the crashes, hopefully.

Long-term I still think it'd be better to push this initialization code
onto the game thread with something more game-loop-y in nature but this
seems like a decent stop-gap.