Fix issues with WM_USER allowing process_events to get called before we intend it to be.
SC:R sets a timer to post WM_USER messages during size/move/menu loops
(like we do with a different message). Inside of its handler, it calls
process_events, which meant that if a resize happened before we started
using that for dispatch, it could get called and cause more crashes.

To fix this, we now ignore WM_USER messages during that initial period.
When we stop ignoring them, we also now properly acquire the event
processing lock to ensure the game memory is properly protected.

This is definitely a mess, hopefully we can clean this up soon :)