Fix issue with lobby init sometimes desyncing.
Our code to set lobby state 8 ("waiting for start") and sending the
start command was pretty close together (actually it was the same
function), which can be bad if users are receiving turns frequently. If
a user receives the start command while *not* in state 8, the contents
of that command are ignored.

Thus, since we receive turns during lobby countdown, if one user is
somewhat delayed from the host, they sometimes received the start
command before they were ready to see it. In this case, their RNG seed
doesn't get initialized, so even though they still start the game on the
correct turn, all of the state is off and they immediately fail a
checksum.

This commit adjusts a few things to fix this and just make things more
consistent:

- All turn receiving during this period is managed through
  LobbyInitCompleter (this lets us have fairly consistent timing for
  sending/receiving/processing turns)
- LobbyInitCompleter tracks its state so that it won't receive extra
  turns once init is complete (regardless of whether that completion
  happens during countdown or in the period after)
- Setting lobby state 8 and sending the start command are now separate
  functions, and we set lobby state 8 much earlier (before countdown
  starts). This is a longer period than the game typically has, but
  shouldn't really be an issue (I'm sure we have all experienced laggy
  games where it takes 5+ seconds to start after the countdown completes
  :) )