Add custom cursor size setting + ability to toggle the cursor size limit patch.
I also moved the mouse scaling setting into a "legacy" area because it's
a dumb setting (love that Blizzard copied my worst ideas from 10 years
ago lol).

The setting to disable the patch is passed via program arguments because
that was an easy way to get it into the game before patching happens,
otherwise we'd have to wait for the websocket and stuff and that seems
non-ideal. So long as there's a limited number of these sorts of
settings I think it seems kind of fine.