Increase the time we consider rally-point pings valid, increase batchability for ping results.
In general it's unlikely network conditions change that much that
quickly, and our previous settings resulted in re-pinging almost every
matchmaking queue. The old settings also resulted in a ton of requests
to report results even though we can batch them, and 300ms vs 66ms will
not really make a difference in us picking routes for people (especially
since the refresh is triggered on queue entry).