Reset saved app window position (but not maximized state).
Probably should have done this for 10.0, but some people have their
windows sized to the old default, and that default can be quite small.
This will force a re-evaluation of the (new) defaults based on screen
size, so hopefully the vast majority of people end up at sizes where the
social sidebar is pinned/pinnable and they can see all the nav items.