Make DumbSpinLock slightly less dumb.
This makes the lock able to be use re-entrantly, which will fix issues
with resizing the window during loading that we were seeing. Namely,
starting the game in a fullscreen mode and hitting alt+enter before the
countdown begins, then moving the window once the countdown starts
seemed to pretty consistently result in deadlocks. It looks like the
issue was in the WM_USER handling by process_events, which was then
triggering our own timer for lobby init progression.

I could have probably found a way to do the proper lock/unlock sequence
for this case as well, but it felt safer to just make our lock
re-entrant, and there are no issues with the same thread
accessing/modifying the game data, so long as it's holding the lock.