Reorganize user profile admin stuff.
Splits the various parts into separate subpages, with netsted navigation
(lol) under thte admin tab to get to them. Fixes #1210.

Also adds a prop to IconButton to let us pass an `as` prop for its root
component (useful when you need a non-button to put inside an anchor
tag). And makes it allow strings for its `icon` prop (which it will pass
to `MaterialIcon`.