Don't render the context menu contents if it doesn't have any items.
Can be reproduced by trying to right-click a chat message in the
whispers session; or for users who don't have the "Delete message"
action, by right-clicking any message in the chat.

This keeps the Popover still open, but figuring out how to close/not
open the Popover in that case seems much harder :d, and I think just not
rendering the Popover contents might be fine?