Make matchmaking accept more resilient to terrible connections.
Clients will now receive the matchmaking accept status over the
subscription for the match (which previously didn't have any initial
data) and it will be updated to the current state (including num
accepted + time left). This will make it so if a client manages to
reconnect before the server realizes they were disconnected, but they
disconnected while in match accept, they'll see the dialog and be able
to accept.