Upgrade JS deps.
Main issue was the TS 5.9 ArrayBuffer changes, not really sure of the
best way of dealing with them, probably I could have just casted to
`any` but didn't feel great. For now we just construct a new UInt8Array
which is definitely a bit inefficient, but it's probably fine for where
it is used.