Remove SbUser dependency from MessageInput for mentionable users. (#1241)
Also, some minor improvements to the mention popover:
- Offline users are now displayed as faded (similar to their UI in the
  channel user list) so it's easier to find online users in the list
- 'Tab' key now selects the focused user

I was also gonna make the first user focused by default when the popover
is open, but that's a bit trickier to do since focused menu items steal
keyboard focus from the message input and we obviously don't want.

The only way to get around that from what I see would be to add a "fake"
focus style to the menu item, which would then have to be controlled by
the message input itself. Or not use our existing MenuList/MenuItem
components and create something custom specifically for this use case.

But I think it works good enough for now. User can still focus the first
menu item by pressing the down arrow, even though it might not be the
most obvious thing in the world.