Remove browserprint stuff.
We weren't really using it anyway, and signup requires the app now
anyway (which has better identifiers than this). I left the existing
ones in the DB but they'll get cleaned up after a few months (and we no
longer store new ones from old clients once this ships). This reduces
some of the download necessary for clients on first load.