Fix message menus not showing up without selecting text.
Broken in 616923056d429ff1c00e71c71ad6b9f52b933152. The MessageMenu is
passed in via props and can be customized/add its own items, so checking
the size there can't work. Those are given `MenuComponent` (which we
pass in `MessageContextMenuList` to actually render their menu, so we
just check the length in there instead.