Update server_rs deps.
I had to set a specific version for tracing-attributes:

cargo update -p tracing-attributes --precise 0.1.28

due to an issue with later versions of that and async_graphql. Hopefully
that will be fixed soon.

ipnetwork continues to be held back for sqlx (they updated the version
of the dep yesterday, but it'll be part of 0.9 which isn't out yet).