Implement login name changes.
These are limited to once per 30 days to limit abuse (tracked via a
`last_login_name_change` added to the users table). There is an
additional audit table that gets updated any time a change happens.
Currently, the only way for a change to happen is via account settings,
but having it available to admins may be helpful (and if/when that
happens, there is also a `reason` column in the audit table to store
that info).

There is a new user admin page (keyed off the same permission as IP
addresses) that lets you view the log of login name changes for a user.
When we add display names, we can put those there too.