Use ReduxMapThumbnail in all places that currently use MapThumbnail.
As part of this work, I also changed what we show on the map thumbnail
in certain places (screenshots below). I used my best judgement for
this, but it can easily be changed.

Also, the MapThumbnail component is a bit weird because it uses one prop
(`showMapName`; false by default) to decide whether to not only show the
map name, but also the map actions. And map actions (view details,
download, etc.) are all visible by default in ReduxMapThumbnail. But
unless you set the `showMapName` to `true`, the actions won't be
visible regardless. Not sure if we would even want to show map actions
without the name, but yeah, something to think about.