Improve error handling when images are over the filesize limit. (#1254)
I went with the option of just adding a custom validator on form
that checks the file size for file inputs. This also means that in order
to show the errors on file inputs, you need to send the `allowErrors`
property to the FileInput component.

And since showing errors on file inputs impacts the UI of the form a
little bit (there is a reserved space for errors), I adjusted the UI a
little bit to better fit those constraints. I also added an optional way
to show the file name in file inputs to fill the space a bit if you need
it.

Finally, while adding the file size errors to the league create/edit
forms, I also made the league preview display the chosen image/badge.

* Update the league API to not send JSON-ified image and badge.

Also adds a missing checkbox to delete the current badge when editing a
league.