Use a multiline textarea element in the MessageInput component. (#1258)
This allows for multiline content to be posted in chat, which I think is
a nice-to-have feature. I also tried emulating Discord's logic where
Shift + Enter is used to enter new lines, and Enter sends the message.

I also tried making the TextField component generic so its event
handlers accept HTMLInputElement or HTMLTextAreaElement based on the
`multiline` prop, but this turned out to be beyond my TypeScript-foo :d

There is a new CSS rule (`field-size: content`) that I used here to have
an auto-sizeable textarea component, instead of using those CSS tricks
we had before (and which don't work when you call `preventDefault` on
textarea). However, this new rule is not supported in Firefox and Safari
yet, so not sure if that's a dealbreaker for us.

And finally, I also changed how we're displaying the text messages in
chat, by preserving whitespace. This allows people to more easily post
stuff like chat logs or ascii memes.