Fix the textarea sizing.
There were a couple of issues here, but the main one was caused by me
not realizing that `field-sizing: content` first expands horizontally
(if there's space), and then it expands vertically only if it can't go
horizontally anymore (e.g. constrained by parent).

This would mean that the textarea could potentially stay indefinitely in
a single-line if it weren't constrained by the parent. Fortunately I
found only a single place where the textarea wasn't stopped from
overflowing, which was when setting the urgent message.

However, the second issue with the previous code was that I added the
`field-sizing: content` not only to textarea element, but to all input
elements as well ><. Which meant that the above issue was also
applicable to all the places where we used the input element. I fixed
that now by only applying this CSS rule to textarea elements.

Finally, I had to make some adjustments in how we apply min-height to
textarea elements now since the `rows` property is not respected when
setting the default height of the textarea (this is by design). And
also, the calculation for max-height was also slightly off since the
paddings were updated recently, but this calculation wasn't updated.