Add buttons for watching/downloading replays from game results page.
For now the permissions are:
- Anyone can view replays from matchmaking
- Only participants can view replays from lobbies

Possibly in the future we'd add privacy settings or the ability to share
a link to one particular replay, but for now we're just taking the
simple path.

Only the longest replay will be shown to normal users, admins with debug
privileges can see every replay uploaded for a game. May need to adjust
this behavior somewhat as it won't work well for desynced games, but
those are probably not that worth watching anyway.